requirements.txt
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2
